refactor : #124 createRoomConnectResponseDto 생성 로직 분리
