+++ src/main/java/com/ssafy/freezetag/domain/room/service/RoomService.java
@@ -12,10 +12,10 @@ import com.ssafy.freezetag.domain.room.entity.Room;
 import com.ssafy.freezetag.domain.room.entity.RoomRedis;
 import com.ssafy.freezetag.domain.room.repository.MemberRoomRepository;
 import com.ssafy.freezetag.domain.room.repository.RoomRepository;
-import com.ssafy.freezetag.domain.room.service.request.OpenviduResponseDto;
+import com.ssafy.freezetag.domain.room.service.response.OpenviduResponseDto;
 import com.ssafy.freezetag.domain.room.service.request.RoomCreateRequestDto;
 import com.ssafy.freezetag.domain.room.service.response.RoomConnectResponseDto;
-import com.ssafy.freezetag.domain.room.service.response.RoomMemberInfoDto;
+import com.ssafy.freezetag.domain.room.service.response.RoomMemberInfoResponseDto;
 import com.ssafy.freezetag.domain.room.service.response.RoomUserJoinEvent;
 import com.ssafy.freezetag.global.util.CodeGenerator;
 import lombok.RequiredArgsConstructor;
@@ -27,6 +27,7 @@ import org.springframework.transaction.annotation.Transactional;
 import java.util.List;
 
 import static com.ssafy.freezetag.domain.room.service.helper.RoomConverter.convertToMemberInfoDtos;
+import static com.ssafy.freezetag.domain.room.service.helper.RoomConverter.createRoomConnectResponseDto;
 import static com.ssafy.freezetag.domain.room.service.helper.RoomValidationHelper.validateDuplicateMember;
 import static com.ssafy.freezetag.domain.room.service.helper.RoomValidationHelper.validateRoomCapacity;
 
@@ -73,8 +74,8 @@ public class RoomService {
         // 현재 엔티티가 LAZY로 설정되어 있어서 이렇게 페치 조인해오지 않으면 쿼리가 여러번 나간다.
         Room room = fetchRoomWithMembers(unfinishedRoom.getId());
 
-        // RoomMemberInfoDto (memberId, memberName) 형태로 변환
-        List<RoomMemberInfoDto> memberInfoDtos = convertToMemberInfoDtos(room.getMemberRooms());
+        // RoomMemberInfoResponseDto (memberId, memberName) 형태로 변환
+        List<RoomMemberInfoResponseDto> memberInfoDtos = convertToMemberInfoDtos(room.getMemberRooms());
 
         // Openvidu 로부터 토큰 및 세션 ID 반환
         OpenviduResponseDto webrtcDto = openviduService.createRoom();
@@ -85,15 +86,8 @@ public class RoomService {
         // Redis에서 채팅(메세지) 정보 불러오기
         List<MessageRedis> messages = messageService.getMessages(room.getId());
 
-        return new RoomConnectResponseDto(
-                room.getId(),
-                enterCode,
-                dto.getRoomName(),
-                dto.getRoomPersonCount(),
-                memberInfoDtos,
-                room.getHost().getId(),
-                webrtcDto,
-                messages);
+        return createRoomConnectResponseDto(room, enterCode, dto.getRoomName(), dto.getRoomPersonCount(), memberInfoDtos, webrtcDto, messages);
+
     }
 
     @Transactional
@@ -124,14 +118,14 @@ public class RoomService {
         memberRoomRepository.save(nowMemberRoom);
         fetchJoinedRoom.getMemberRooms().add(nowMemberRoom);
 
-        // RoomMemberInfoDto (memberId, memberName) 형태로 변환
-        List<RoomMemberInfoDto> memberInfoDtos = convertToMemberInfoDtos(fetchJoinedRoom.getMemberRooms());
+        // RoomMemberInfoResponseDto (memberId, memberName) 형태로 변환
+        List<RoomMemberInfoResponseDto> memberInfoDtos = convertToMemberInfoDtos(fetchJoinedRoom.getMemberRooms());
 
         // OpenVidu 토큰 반환
         OpenviduResponseDto webrtcDto = openviduService.enterRoom(sessionId);
 
         // 신규 회원 정보를 event 형태로 저장
-        RoomMemberInfoDto newMemberInfo = new RoomMemberInfoDto(memberId, member.getMemberName());
+        RoomMemberInfoResponseDto newMemberInfo = new RoomMemberInfoResponseDto(memberId, member.getMemberName());
         RoomUserJoinEvent roomUserJoinEvent = new RoomUserJoinEvent("NEW_USER_JOINED", roomId, newMemberInfo);
 
         // 기존 사용자에게 새로 입장된 유저 알림
@@ -141,15 +135,8 @@ public class RoomService {
         List<MessageRedis> messages = messageService.getMessages(roomId);
 
         // 현재 방 정보 반환
-        return new RoomConnectResponseDto(
-                roomId,
-                enterCode,
-                fetchJoinedRoom.getRoomName(),
-                fetchJoinedRoom.getRoomPersonCount(),
-                memberInfoDtos,
-                fetchJoinedRoom.getHost().getId(),
-                webrtcDto,
-                messages);
+        return createRoomConnectResponseDto(fetchJoinedRoom, enterCode, fetchJoinedRoom.getRoomName(), fetchJoinedRoom.getRoomPersonCount(), memberInfoDtos, webrtcDto, messages);
+
     }
 
     public Room fetchRoomWithMembers(Long roomId) {
