+++ src/main/java/com/ssafy/freezetag/domain/room/service/helper/RoomValidationHelper.java
@@ -5,7 +5,7 @@ import com.fasterxml.jackson.databind.ObjectMapper;
 import com.ssafy.freezetag.domain.exception.custom.DuplicateRoomMemberException;
 import com.ssafy.freezetag.domain.exception.custom.RoomFullException;
 import com.ssafy.freezetag.domain.room.entity.Room;
-import com.ssafy.freezetag.domain.room.service.response.RoomMemberInfoDto;
+import com.ssafy.freezetag.domain.room.service.response.RoomMemberInfoResponseDto;
 import com.ssafy.freezetag.domain.room.service.response.RoomUserJoinEvent;
 import org.springframework.messaging.simp.SimpMessagingTemplate;
 
@@ -23,7 +23,7 @@ public class RoomValidationHelper {
     }
 
     public void sendNewUserJoinEvent(Long roomId, Long memberId, String memberName, SimpMessagingTemplate messagingTemplate, ObjectMapper objectMapper) throws JsonProcessingException {
-        RoomMemberInfoDto newMemberInfo = new RoomMemberInfoDto(memberId, memberName);
+        RoomMemberInfoResponseDto newMemberInfo = new RoomMemberInfoResponseDto(memberId, memberName);
         RoomUserJoinEvent roomUserJoinEvent = new RoomUserJoinEvent("NEW_USER_JOINED", roomId, newMemberInfo);
         String joinEvent = objectMapper.writeValueAsString(roomUserJoinEvent);
         messagingTemplate.convertAndSend("/sub/room/" + roomId, joinEvent);
